refs codership/galera#367 use system ASIO if available
Made debian build to depend on libasio-dev and enhanced SConstruct
to check if system provides ASIO devel library and if it is recent
enough (> 1.10.1).

If ASIO devel library is not installed or is not recent enough,
build falls back to use of embedded copy.

Also reduced the scope of compiler warning suppressions when
system ASIO is used and fixed compiler warnings.