- PXC #384: Garb init script causes new UUIDs to be generated every time.
This error occurs when the configuration file is not specifies base directory (there is no base_dir option) AND garbd do not have write access to the current directory. For example, if it is running as a daemon, the current directory can be / and garbd unable to write the ./gvwstate.dat file. After this patch, garbd makes some tests:
- whether the current directory is root (/) or not?
- whether we have write access to the default directory (./) or not?
If the current directory is root or we do not have write access to the current (default) directory, then patch thinks that we are working as a daemon.
Then it tries to use the /var/lib/galera/ path to store gvwstate.dat [G in a similar way, as so many other linux demons do. If the /var/lib/galera directory does not exist, garbd trying to create it. If it is impossible, garbd signals a fatal error.