This is patch for the https://bugs.launchpad.net/percona-xtradb-cluster/+bug/1532857 issue.
In the past, we are faced with the problem that the Galera’s cluster arbitrator daemon (garbd) does not keep uuid when the user restart it (https://jira.percona.com/browse/PXC-384).

This error occurs when the configuration file does not specified base directory and arbitrator daemon does not have write access to the current directory (because default base directory = "./"). For example, if it is running as a daemon under "nobody" user, the current directory can be "/" and garbd is unable to write state file (without root privileges). Therefore, after restart, it cannot read the uuid from a nonexistent state file.

After PXC-384 patch, garbd daemon makes some tests and if the current directory is global root ("/"), unreachable or garbd does not have write access to the default directory ("./"), then garbd thinks that it launched as a daemon. Then it tries to use the "/var/lib/galera/" path to store state file – in a similar way, as so many other linux demons do. If the "/var/lib/galera" directory does not exist, the garbd trying to create it. If it is impossible, then garbd is signaling a fatal error.

However, because the PXC installer does not create this directory yourself, we have a problem.

On the other hand, without the PXC-384 patch, we would have improperly working garbd, which is unable to read own state after a restart.

Therefore, we need to create a "/var/lib/galera" directory in the installer, and assign it to "nobody" user.

This patch adds this directory.