Seqno in the grastate.dat file is reset to "-1" (although the state is not marked as "unsafe") each time when the server is started (this done in Galera initialization, see the galera::ReplicatorSMM::update_state_uuid() function) and remains undefined until completion of the SST or IST.
To correct this error is sufficient to pass the current value of seqno into st_.set(...) function that is called from the galera::ReplicatorSMM::update_state_uuid(). Currently it called with the WSREP_SEQNO_UNDEFINED argument that causes this bug.

We also need to add a new logic into the recv_IST() function, to keep the state marked as having an undefined position after the start of the data changes during IST. To do this, if the current position is defined (for example, when there were no SST before IST), then we need to change it to an undefined position before applying the first transaction, since during the application of transactions (or after the IST) server may fail.