PXC-553 : Crash when setting fc_limit=20000 and then running stress
Issue:
Set the fc_limit=20000, and then block the node ("flush tables with read lock").
The queue will eventually reach the limit (which is actually ~34800), but
before that it will reach the size of the receive queue (which is set upon
startup, and on a 256MB machine will hold 32K entries). If you then
"unlock tables", this will hit an assert in the code.

There is a bug in the fifo queue implementation that if the circular buffer
has wrapped (so that -->tail ... head-->) and that head and tail are in the
same memory block, when head reaches the end of the block,it will free()
the block, thus erasing part of the entries.

Solution:
(1) Fix the bug in the fifo queue so that the block is not prematurely released.
(2) Ensure that the limits resulting by setting fc_limit do not exceed the
    maximum size of the queue.