GCF-933 Fixed the condition for primary component recovery
The condition used std::equal() to check member set equality without
checking that the member sets were of equal size. Therefore if the
delivered view size was smaller than the rst_view size and the
delivered view member set was equal to the prefix of the rst_view
member set, the primary component was bootstrapped erroneously.

The fix was to check that the member sets of the delivered view
and the rst_view are of equal size.