GCF-942 Introduced safe_to_bootstrap flag
When restarting the whole cluster it is important to shut down the
nodes in the opposite order than they will be started to make sure
that the node which was shut down last is started the first.
Even if there is no load present this is necessary to keep
view information delivered to application consistent accross
cluster restarts.

This patch introduces a safe_to_bootstrap flag which prevents
bootstrapping a new cluster if the node to be used to bootstrap
a new cluster was previously a member of a primary component
and was not the last node to be shut down.

Safe_to_bootstrap flag is stored into grastate.dat file
and can be reset either by editing the file manually or
by removing the file.