GCF-980 Check trx->state() while holding its lock in galera_release()
This fixes a race that happens if galera_abort_pre_commit() is called
on a trx handle that is being released. In which case it is possible
for galera_abort_pre_commit() to set the state of the trx to S_MUST_ABORT
and trigger an assertion in the sanity checks of galera_release().
To solve the issue we move the sanity checks to happen while the handle
is locked.