GCF-895 Reset slave transaction pointer when assigning a new one
The slave transaction pointer was released in TrxHandleMaster::finalize(),
but it is too early. If the send fails for some reason and it will
be retried, old slave transaction context will be missing.

Postpone releasing slave transaction to TrxHandleMaster::add_replicated().

This commit contains also change to truncate transaction state history
whenever S_EXECUTING is reached.