GCF-1016 Fix assertion ref_trx != trx in certify_and_depend_v3
Assertion ref_trx != trx would trigger because trx was certified
against an old trx that was not purged from cert_index, both
shared the same address and had a key in common.
The old trx was not purged because it was marked as 'dummy' and
its depends_seqno reset to -1, even though the transaction did
pass certification.
The fix is to avoid resetting depends_seqno for those transactions
that passed certification.