GCF-1018 Fix hang due to aborted pre_commit not leaving apply monitor
This patch fixes a hang that is due to ReplicatorSMM::pre_commit()
not leaving apply_monitor on bf abort. Specifically, this fixes
a case where a trx in pre_commit() managed to grab the apply monitor,
but is BF aborted before it manages to lock trx.
The fix is to check whether monitor was entered before BF abort, and
if so leave it.