GAL-501 Improved URI parsing for IPv6 addresses
Method gu::URI::parse_authority() could not handle the case where the
URI consists of an IPv6 address without a port.
The implementation happened to parse [::1]:789 correctly, but it would
fail to parse ::1 or [::1].
Addresses that start with '[', or contain two or more ':', are assumed
to be IPv6 addresses and parsed accordingly.