GCF-1043 Fix certification index divergence
This patch fixes the case where the certification index could
diverge due to BF aborted transactions that are not going to
replay. More precisely, streaming transactions that are not
committing (F_COMMIT flag unset), and are BF aborted before
certification, could be delivered and certified successfully
by slaves. As a result, the certfication index of master and
slaves could diverge.
To avoid this case, we push those writesets in a queue (see
PendingCertQueue), which is ordered by their seqno, and make
sure they are certified in order.