GAL-518 Improve AsioTcpSocket::write_handler() error handling
If any of the sanity checks for the write_handler() failed, the
exception was thrown and the gcomm backend terminated.

This patch improves the error handling in such a situations by
reporting a protocol error for upper layers and closing the socket.
The socket will be reopened once the upper layer checks the
connections status.