GAL-515 "scons ssl=0" does not disable all SSL dependencies
"scons ssl=0 ..." does not compile if SSL is not present on the system.
"Fix" it by removing that option and unconditionally require SSL to be
present in order to compile Galera. Still, at runtime, Galera may be
configured to use plain (non-SSL) connections.

"Fixing" it in this way is mandaded by the complexity of supporting
non-SSL builds and handling gracefully MySQL configs like
"wsrep_provider_options=...;socket.ssl=yes;...". In order to handle that,
the Galera library has to proclaim to MySQL that it does not support SSL,
so that it can return some meaningful error message to the user and refuse
to start.