Define HAVE_STD_* macros for the C++11 structs that we use
HAVE_STD_foo denotes that we are compiling in C++11 mode and that "foo"
is present and can be used ("foo" is some C++11 feature).

HAVE_STD_SHARED_PTR was used instead of HAVE_STD_MEMORY because the
<memory> header is present in pre-C++11 environments.

This patch also assumes that if std::unordered_map is present, then
also unordered_set, hash, unordered_multimap and unordered_multiset are
present. This is fine, because, although implicit, we know that all
those are present in C++11 mode (and HAVE_STD_UNORDERED_MAP denotes that
we are in C++11 mode).

Suggested by: ayurchen