PXC-911 : "no messages seen in" from itself when starting
Issue:
A message is set to the info log when no messages are received from a node
for a certain timeout.  However, in this case, the address has been
blacklisted since it's coming from itself.  Becasue it's blacklisted,
we would not expect to hear from that address anymore.

Solution
Only log the message if the node is NOT blacklisted.