Refs codership/galera-bugs#111: valgrind complains about uninitialized memory
This is caused by using uninitialized memory for padding the recordset to a
multiple of alignment size. This uninitialized memory is then checksummed
and valgrind is concerned about the checksum value. However in this case
we don't care about the value of the checksummed data, all we are concerned
with is that the data remains unaltered through copying and transmission.
Added padding bytes initialization for debug builds.
Also fixed a harmless memory leak complaint in gu_net_test.cpp to ensure a
clean valgrind run.