Avoid unnecessary GCache reset after recovery
GCache is sometimes dropped unnecessarily right after it was recovered.
This happens when the highest sequence number in the recovered GCache
does not match the recovery sequence number provided by the application
upon startup. It is possible that GCache has higher maximum sequence
number, meaning that GCache recovered more events with respect to the
application.

This patch changes method GCache::seqno_reset() to detect this situation
and adds method GCache::discard_tail() to discard GCache entries with
seqno greater than the given seqno. Instead of resetting the whole
GCache, only those entries are discared which have not been recovered
in the corresponding application state.

codership/galera#498