Fix posix_fallocate() error handling
From posix_fallocate(2):

  If successful, posix_fallocate() returns zero. It returns an error on
  failure, without setting errno.

So, if that function fails, don't expect errno to be set. This is important
because we have special handling if `EINVAL == errno || ENOSYS == errno`.
This fixes an error at startup:

  2018-03-12 15:35:24 34391166976 [ERROR] WSREP: File preallocation failed: 0 (No error: 0) at /home/

Also, because fcntl()'s specification says it returns -1 on failure, check for
failure by `== -1` rather than `!= 0`. It could return 0 on success.