codership/galera#500 Properly assing leave comp msg size
The problem was that if the gcomm background thread terminated
via exception, the message size for leave comp message was not
set appropriately and the message got discarded by gcs. This
lead to partial shut down where replicator, gcs were left operational
but gcomm was shut down.

The fix was to assing leave comp message size.