codership/galera#501 Added cluster weight to status variables
Added a new status variable which shows total cluster weight,
computed as a sum of node weights as configured with pc.weight.

The total cluster weight is computed for current primary component
only, not including weight of nodes which have been partitioned
or crashed. Cluster weight in non-primary component is always zero.