codership/galera#511 SSL_CTX_set_ecdh_auto() needs to be called explicitly.
Some older versions of OpenSSL require SSL_CTX_set_ecdh_auto() to be
called explicitly in order to enable advanced algorithms. Without
that the server may refuse connection citing 'no shared cipher'.
This commit checks for the availablity of that method at compile time,
and if available a call is compiled in.