Changes for XA recovery
Allow fragments that terminate a transaction (either commit or
rollback) to be replicated even if galera does not have a trx handle
for those transactions. This is useful to terminate xa transactions in
prepared state.
This achieved by changing galera_certify() to take trx meta as in
parameter, so that any node can terminate a prepared transaction on
behalf of another node.