Implemented some extra debugging instrumentation.
In particular, added `gcache.debug` integer parameter available
in debug builds which is a bitmask to turn on debug logging in
different GCache modules, e.g.
  1 - for MemStore,
  2 - for RingBuffer
  4 - for PageStore
and the rest of the code at the moment just looks if it is non-0.

Since the performance impact of this is unclear, the code is
wrapped in `#ifndef NDEBUG` to compile only in debug builds.