Refs codership/galera-bugs#201 seqno_released is advanced too optimistically
GCache::free_common() method advanced seqno_released pointer too
optimistically in the case of buffers stored in PageStore. Since
GCache pages are volatile, freeing (i.e. releasing) a buffer may
actually mean page deletion at any time so it is equivalent to
discarding (disposing of) the buffer (and seqno) altogether. And
so we must make sure that all preceding seqnos/buffers can be
discarded. So we actually need to track the highest seqno that was
possible to discard and consider it the highest released.