Enabling C++11 compilation for C++ compilation units
Flag -std=c++11 is added to CXXFLAGS, or fallback -std=c++0x is
used if the compiler does not understand c++11. This is the case
with GCC 4.4.

When deciding which shared_ptr implementation to use, prefer
boost implementation since it currently works the best with
code which mixes asio and boost.

Fixed type deduction in TrxHandleSlave to make it work also with
older compilers.