codership/galera#533 Prevent nodes with the same UUID in the cluster
Implemented check in gmcast for connections from other nodes which
have the same UUID. Abort gcomm background thread if a node
with the same UUID is detected and the primary view has not
been reached. This will make the joining node to abort, but the
node which has reached a primary view at least once will
continue.