Refs codership/galera#534 Unescape IPv6 bind_ip when joining the cluster
When IPv6 is used in option gmcast.listen_addr, it must be enclosed
in square brackets. This option is also used as the source address to
bind to before connecting to the cluster.

The source address cannot be bound to an IPv6 address when it contains
square brackets, so make sure we removed them when needed.