codership/galera-bugs#331 Clear trx meta GTID on TOI error
Trx meta is given as an in/out parameter to galera_to_execute_start().
In case of NBO end, the meta will contain the GTID of the corresponding
NBO begin. If the TOI cannot be grabbed for the NBO end due to closed
provider or communication error, the meta may still contain the
original GTID when the function returns.

As a fix, clear the meta GTID part if there is no valid GTID
associated to attempted TOI operation.