codership/galera#553 Ignore ENOBUFS error for keepalives.
Keepalive messages are used to signal other when the socket
is found idle. However, if the send operation returns ENOBUFS,
it means that the socket has some data to send in outbound queue
and the keepalive error can be ignored. If the data cannot be
transmitted, the connection will fail anyway.