codership/galera¤554 Request retransmission only for missing messages.
The EVS retransmission protocol was too greedy. The retransmission
was requested for all messages starting from the first missing message
up to highest seen. This caused excessive flooding of network with
large EVS send window sizes.

As a fix, compute a list of ranges for missing messages and send
separate gap message to request retransmission for each range
separately.