Remove obsolete target 'syslog' from systemd service definition
Declaring After=syslog.target is unnecessary by now because syslog is
socket-activated and will therefore be started when needed.

Having network.target is enough, and that is e.g. how the MariaDB systemd
service defines it.

This change is safe on the Galera 4.x branch since it most likely will not
target very old Linux distros anymore.

See also:
https://lintian.debian.org/tags/systemd-service-file-refers-to-obsolete-target.html