codership/galera-bugs#379 EVS representative missing own prev incarnation
If a new incarnation of the node becomes a representative after fast
restart, the install message it sends to the group will not contain
the old incarnation.

Work around this by discarding messages from the new incarnation
until a new view which does not contain the old incarnation has
been installed.