Fix systemd unit and start script
The start script does not work on Debian because the path for config
file is not correct.

Also, the bash script needed some linting IMO.

Finally, the environment definition in the systemd unit file is
redundant (and problematic on Debian) with config file definition in the
bash start script (garb-systemd).