Set base_dir by default as /var/lib/galera instead of / (Closes: #313)
The old default value '.' often translate simply to system root '/' when
garbd (most often) is run from a system init/service file that has no
path or is just '/' to begin with. This is a bad default, garbd should not
imply that it wants to write to file system root.

Many users seem to set this manually to /var/lib/galera or /var/lib/garb.
On Debian/Ubuntu systems I've seen users set in /etc/default/garbd:
  GALERA_OPTIONS="base_dir=/var/lib/galera/"

This change in the global default will render such hacks unnecessary.