Set depends_seqno on dummy writesets
This patch makes changes to certification so that we preserve the
depends_seqno of writesets that fail the certification test.
This change avoids the case where we deliver such a writeset to
the DBMS, which is still the process of BF aborting the same
transaction.
This may happen in the case of streaming replication, where it is
possible that the transactions manages to replicate one more fragment
before BF abort, and the fragment fails certification.