codership/galera#606 Don't print TrxHandleSlave state history
Currently there is no sufficient locking for TrxHandleSlave
to allow concurrent access to mutable data in the object.

In order to avoid printing garbage when logging certification
conflicts, avoid printing state history that is stored in
dynamically allocated vector for TrxHandleSlave. The state histry
for TrxHhandleSlave is mostly uninteresting anyway because the
state graph is linear.

Printing state history for TrxHandleMaster remains.

codership/galera#605 Added static_cast<int> as suggested.