Don't attempt to release trx for ws handle with null opaque
When trx handle is created for a transaction, it is assigned
to wsrep_ws_handle_t::opaque. Don't try to release a trx object
for ws handle which has opaque set to NULL, as there has not
been active transaction for that handle and the associated
trx_id may not be unique.

This is a defensive mechanism for application which does not
track the provider side transaction lifetime properly.