Always allow sending join messages when not in JOINER state
Relaxed condition is required for correct handling of desync counter,
especially in DONOR state: If the DONOR does desync in combination
with SST donation, the gcs_join() calls from resync() and sst_sent() might
come with out of order seqnos, leaving the desync_count in gcs_group
permanently in non-zero value. In this case the node will not become
synced again unless it is temporarily removed from the group.

Refs: MDEV-27459