Fix setting of invalid option values
Galera library accepts wrong parameter values as long as the value has
a "good" prefix. For example, parameter `evs.log_info_mask` is parsed
as an integer using hex notation such as `0x1`. However, a good prefix
was enough for the parsing to accept a value, so attempting to set it
to  `0x1whatever` resulted in value `0x1`.
This patch fixes `gu::from_string()` so that it raises an exception if
it detects such case. Notice that the `gu::from_string<bool>`
specialization was already catching such cases.