Race between action commit and commit cut applying: codership/galera-bugs#901
Non-SYNCED nodes are not counted when calculating commit cut - maximum
seqno that is no longer needed in certification index (all nodes have
committed it). As a result a situation may arise when group commit cut
is applied before all preceding events are committed on a non-SYNCED
node. In such case drain all monitors to ensure that all preceding
events are committed.