garbd fails to exit gracefully in case GCS layer throws an exception.
In fact the process needs to be killed with an abort signal. The reason
is that the exception is not caught inside the loop which makes control
to exit the receive loop and close GCS connection AFTER that. As a result
the receive queue cannot be emptied and the process hangs. Thus any error
in the configuration may result the hanging process.

Make sure all exceptions are caught inside the receive loop and cause
connection close first. And after that the loop loops until SELF-LEAVE
message is delivered.

Refs codership/galera-bugs#931