Refs codership/galera-bugs#938 Commit cut tracking is broken on node leave
Suppose we have a 3-node cluster and node 3 is marked as having the
smallest LAST_APPLIED seqno. Then if node 3 leaves the cluster, leaving
nodes 1 and 2 in primary view, LAST_APPLIED messages from the remaining
nodes no longer will be counted and commit cut won't progress. As a result
certification index won't be purged in time. This happens because GCS
would still be waiting for LAST_APPLIED from node 3 because
group_redo_last_applied() is not called on new primary view.

This requires bumping GCS protocol version for backward compatibility.