Run gcomm join process in gcomm service thread
Running gcomm connect in the thread that is executing the join
process may cause issues if the allowlist callback is called
and the thread is DBMS server thread that is handling client
connections. The allowlist calls are assumed to happen solely
from internal Galera service threads.

As a fix, change the gcomm join process to happen in gcomm
service thread.