Add application defined sequential consistency for certification
In order to provide means to implement application defined
sequential consistency for certification, the wsrep-API is
updated to version which has an extension wsrep_certify_v1().

The extension call will take an application defined callback
as an argument. The callback is called once the sequential
consistency is guaranteed during write set replication.