Set accepted socket in S_CONNECTING state
The connection cleanup did not happen if an accepted connection
was closed during handshake. The reason was that the socket
state was not set to S_CONNECTING appropriately, and cleanup
was not complete for the socket because it was still in S_CLOSED
state.

Fix by setting the socket state in S_CONNECTING once accepted.