Update wsrep API and implement config service v2
This implements version 2 of the config service. Version 2
of the service maintains the same interface and functionality
of version 1. The difference is that version 2 is able to
retrieve the complete list of parameters right after the
Galera library is loaded, before initializing the replicator.
This commit also changes the default initialization of some,
parameters, i.e. from constructor to static initialization
were possible.