MGL-75 Old last_applied value is inherited into new history
If a node rejoins a new cluster component with a new history (for example
a new primary component is bootstrapped during network partition) it still
carries old and likely very high last_applied counter with it and

 - will be unable to process new LAST messages from itself as long as their
   seqno is below that last_applied value preserved from old history.
 - will "infect" the rest of the cluster with that last_applied value if it
   for any reason becomes the only member of the primary configuration.

It is a regression and was introduced in commit ddbcc2917d.

Revert that mimstake (requires GCS protocol bump) and implement a way to
"desinfect" the cluster.