MDEV-22124 Rogue JOIN message prematurely joins a donor node
Followup to 065e484144c5999709ae8fd19844da72bb785073

There is a race between an applier thread that sends JOIN after SST is
complete and the GCS recv thread that switches Galera states and also
can send JOIN message in case of a configuration change prevented the
applier thread to do so. However the flag that regulates that is set
prematurely and allows a race between CC processing by the recv thread and
JOIN sending by the applier thread. Synchronize access to this flag and
set it to true only if sending of the JOIN message truly failed.